Extend the config class to support a history file
The config class returns via history() the path to the history file
When a history file does not exist or is longer than the history-length
setting allows, a new history file is created.

save_settings() and get_settings() now uses also the config class to
check if the config file does already exist.