find_submodules now adds submodules to the config
Via save_settings the submodules are now stored in the config.yml file
I also added a verbose option, by default this value is set to False, so
the error messages are show to the user by default. the behaviour to
show submodules is only relevant, when the user adds submodules manuel,
when the find_submodules() function, which is executed in another thread
win the background the error messages would be very annoying to the
user.